DiffuseST_rl/requirements.txt
accelerate==1.0.1
diffusers==0.29.2
huggingface-hub==0.29.3
numpy==1.24.3
pillow==10.4.0
tokenizers==0.13.3
torch==1.11.0
torchaudio==0.11.0
torchvision==0.12.0
tqdm==4.67.1
transformers==4.33.2
